refactor(audit): split the audit middleware into one file per resource
The middleware had grown to a single 1,212-line file ordered by when each
coverage PR landed, so a resource's specs were scattered across it: `function`
spanned 362 lines, `api_key` 252. Now each audit resource has one file holding
every emitter for it, with the engine, the pre-validation request readers and
the DB lookups in their own files.

auditAuth and auditSyncCommand fold into the appAuth and sync resource files,
so a resource's coverage is never split by whether it is emitted from a spec or
from hand-rolled middleware. Tests move next to the middleware they cover and
share their scaffolding. A layout test pins the arrangement.

No behaviour change: normalising imports and the exports the split had to add,
the production code goes 1,132 lines to 1,130 — the delta is a duplicated
getLogger call — and comment lines are unchanged at 141.
