--- packages/server/lib/middleware/audit.middleware.ts
@@ -1,1179 +1,0 @@
-import db from '@nangohq/database';
-import { accountService, configService, customerKeyService, environmentService, getInvitation, getPlanSafe, userService } from '@nangohq/shared';
-import { getLogger, metrics } from '@nangohq/utils';
-
-import {
-    auditEventDropped,
-    changedFields,
-    connectSessionActor,
-    makeAuditTarget as makeTarget,
-    recordAuditEvent,
-    toAuditId as toId,
-    UNKNOWN_ACTOR
-} from '../audit.js';
-import { normalizeSyncParams } from '../controllers/sync/helpers.js';
-import { auditExportQuery, auditListQuery } from '../controllers/v1/audit-trail/query.js';
-import { connectionCreatedActor } from '../hooks/auditConnection.js';
-import { canRecordAuditTrail } from '../utils/auditTrail.js';
-
-import type { RequestLocals } from '../utils/express.js';
-import type {
-    AuditActor,
-    AuditAttribution,
-    AuditContext,
-    AuditEvent,
-    AuditOutcome,
-    AuditTarget,
-    AuditTargetType,
-    AuditVia,
-    MfaVerifiedMetadata
-} from '@nangohq/audit';
-import type {
-    AcceptInvite,
-    AuditActionOf,
-    AuditPolicy,
-    AuditResource,
-    AuditScope,
-    CreateAccountApiKey,
-    CreateApiKey,
-    DeclineInvite,
-    DeleteAccountApiKey,
-    DeleteApiKey,
-    DeleteConnection,
-    DeleteEnvironment,
-    DeleteIntegration,
-    DeleteIntegrationFunction,
-    DeleteInvite,
-    DeleteMFA,
-    DeletePublicApiKey,
-    DeletePublicConnection,
-    DeletePublicEnvironment,
-    DeletePublicIntegration,
-    DeletePublicIntegrationFunction,
-    DeleteSpendAlert,
-    DeleteStripePayment,
-    DeleteSyncVariant,
-    DeleteTeamUser,
-    Endpoint,
-    GetAuditTrail,
-    GetAuditTrailExport,
-    PatchApiKey,
-    PatchConnection,
-    PatchEnvironment,
-    PatchFlowDisable,
-    PatchFlowEnable,
-    PatchFlowFrequency,
-    PatchIntegration,
-    PatchPublicConnection,
-    PatchPublicIntegration,
-    PatchTeamUser,
-    PatchUser,
-    PatchWebhook,
-    PostConnectionMetadata,
-    PostConnectionRefresh,
-    PostDeploy,
-    PostEnvironment,
-    PostEnvironmentVariables,
-    PostFunctionDeployment,
-    PostFunctionDeploymentBundle,
-    PostIntegration,
-    PostInvite,
-    PostMFAActivation,
-    PostMFAEnrollment,
-    PostMFALoginVerification,
-    PostMFARecoveryCodes,
-    PostPlanChange,
-    PostPlanExtendTrial,
-    PostPreBuiltDeploy,
-    PostPublicApiKey,
-    PostPublicEnvironment,
-    PostPublicIntegration,
-    PostPublicQuickstartIntegration,
-    PostPublicRotateWebhookSigningKey,
-    PostPublicSyncPause,
-    PostPublicSyncStart,
-    PostRotateWebhookSigningKey,
-    PostStripeCollectPayment,
-    PostSyncVariant,
-    PutBillingInvoicingDetails,
-    PutPublicSyncConnectionFrequency,
-    PutSpendAlert,
-    PutTeam,
-    PutUpgradePreBuiltFlow,
-    PutUserPassword
-} from '@nangohq/types';
-import type { Request, RequestHandler, Response } from 'express';
-
-const logger = getLogger('Audit');
-
-type AuditRequest<TEndpoint extends Endpoint<any>> = Request<TEndpoint['Params'], TEndpoint['Reply'], TEndpoint['Body'], TEndpoint['Querystring']>;
-
-type AuditableEndpoint = Endpoint<any> & { Audit: AuditPolicy };
-
-const Audit = {
-    auditable: <R extends AuditResource, A extends AuditActionOf<R>, S extends AuditScope>(policy: {
-        resource: R;
-        action: A;
-        scope: S;
-    }): AuditPolicy<R, A, S> => ({
-        kind: 'audit',
-        ...policy
-    })
-};
-
-// Metadata is loosely typed here; per-event shapes live on the emit model (@nangohq/audit's AuditEvent).
-type AuditSpec<TEndpoint extends AuditableEndpoint> = {
-    policy: TEndpoint['Audit'];
-    target?: (
-        req: AuditRequest<TEndpoint>,
-        locals: Partial<RequestLocals>
-    ) => AuditTarget | AuditTarget[] | undefined | Promise<AuditTarget | AuditTarget[] | undefined>;
-    // Created resources expose their id only in the response body — resolve the target from it at finish.
-    // Runs only when `target` produced nothing, so a request-derived target always wins.
-    targetFromResponse?: (
-        response: TEndpoint['Success'],
-        req: AuditRequest<TEndpoint>,
-        locals: Partial<RequestLocals>
-    ) => AuditTarget | AuditTarget[] | undefined | Promise<AuditTarget | AuditTarget[] | undefined>;
-    metadata?: (
-        req: AuditRequest<TEndpoint>,
-        locals: Partial<RequestLocals>
-    ) => Record<string, unknown> | undefined | Promise<Record<string, unknown> | undefined>;
-    // Values known only after the handler responds (e.g. persisted scopes). Merged over request metadata at finish.
-    metadataFromResponse?: (
-        response: TEndpoint['Success'],
-        req: AuditRequest<TEndpoint>,
-        locals: Partial<RequestLocals>
-    ) => Record<string, unknown> | undefined | Promise<Record<string, unknown> | undefined>;
-    // Defaults to the authenticated account (res.locals.account). Override when the audited account is not
-    // the caller's — e.g. accepting/declining an invite is recorded under the inviting team, not the invitee.
-    account?: (req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => Promise<{ id: number; uuid: string } | undefined>;
-    environment?: (
-        req: AuditRequest<TEndpoint>,
-        locals: Partial<RequestLocals>
-    ) => Promise<{ id: number; name: string } | null> | { id: number; name: string } | null;
-};
-
-function omitUndefined(obj: Record<string, unknown>): Record<string, unknown> | undefined {
-    const out: Record<string, unknown> = {};
-    for (const [key, value] of Object.entries(obj)) {
-        if (value !== undefined) {
-            out[key] = value;
-        }
-    }
-    return Object.keys(out).length > 0 ? out : undefined;
-}
-
-export function resolveActor(locals: Partial<RequestLocals>): AuditActor {
-    if (locals.authType === 'secretKey') {
-        // Functions currently call the API with a secret key too, distinguished only by the
-        // client-settable Nango-Is-Script header — spoofable, so we don't trust it for attribution.
-        // Every secret-key caller is classified as api_key until functions get their own tokens.
-        return {
-            type: 'api_key',
-            id: locals.apiKeyId != null ? String(locals.apiKeyId) : 'secret_key',
-            ...(locals.apiKeyDisplayName ? { display: locals.apiKeyDisplayName } : {})
-        };
-    }
-    // An end user is optional when the session carries tags, so the session can name nobody.
-    if (locals.authType === 'connectSession') {
-        return connectSessionActor(locals.endUser);
-    }
-    // Basic auth and auth-disabled deployments authenticate a dashboard user without calling it a session.
-    if (locals.user) {
-        return { type: 'user', id: String(locals.user.id), display: locals.user.email };
-    }
-    // Includes the deprecated public-key flow: it identifies an environment, never a person.
-    return UNKNOWN_ACTOR;
-}
-
-// Never on the impersonating account's own trail: "Nango acted via Nango" says nothing. Events for any
-// other account keep the mark — the request did arrive through that session.
-function auditVia(req: Request, accountId: number): AuditVia[] | undefined {
-    const by = req.session?.impersonatedBy;
-    if (!by || by.accountId === accountId) {
-        return undefined;
-    }
-    return [{ type: 'impersonation', id: String(by.accountId), display: by.accountName, ...(by.actorId ? { actorId: String(by.actorId) } : {}) }];
-}
-
-/** The event fields that are purely a function of the request, so a new one lands on every emitter at once. */
-export function auditRequestFields(req: Request, accountId: number): { context: AuditContext; via?: AuditVia[] } {
-    const via = auditVia(req, accountId);
-    return { context: contextFromRequest(req), ...(via ? { via } : {}) };
-}
-
-export function contextFromRequest(req: Request): AuditContext {
-    const context: AuditContext = { interface: 'api' };
-    if (req.ip) {
-        context.ip = req.ip;
-    }
-    const userAgent = req.get('user-agent');
-    if (userAgent) {
-        context.userAgent = userAgent;
-    }
-    return context;
-}
-
-export function outcomeFromStatus(status: number): AuditOutcome {
-    if (status < 300) {
-        return 'success';
-    }
-    if (status === 401 || status === 403) {
-        return 'denied';
-    }
-    return 'failure';
-}
-
-/** The event still records; the named field is what it lost. */
-export function auditEnrichmentFailed(field: 'target' | 'metadata' | 'display' | 'environment', resource: string, err: unknown): void {
-    logger.warning(`audit event enrichment failed`, { field, resource, err });
-    metrics.increment(metrics.Types.AUDIT_EVENT_ENRICHMENT_FAILED, 1, { field, resource });
-}
-
-// Low-RPS events only — never call this on a hot path (get-credentials derives displays from the request).
-async function resolveDisplay(target: AuditTargetType, lookup: () => Promise<string | undefined>): Promise<string | undefined> {
-    try {
-        return await lookup();
-    } catch (err) {
-        auditEnrichmentFailed('display', target, err);
-        return undefined;
-    }
-}
-
-async function resolveEnvironment<TEndpoint extends AuditableEndpoint>(
-    resolve: NonNullable<AuditSpec<TEndpoint>['environment']>,
-    req: Request,
-    locals: Partial<RequestLocals>,
-    resource: string
-): Promise<{ id: number; name: string } | undefined> {
-    try {
-        return (await resolve(req as AuditRequest<TEndpoint>, locals)) ?? undefined;
-    } catch (err) {
-        auditEnrichmentFailed('environment', resource, err);
-        return undefined;
-    }
-}
-
-async function emit(
-    policy: AuditPolicy,
-    req: Request,
-    res: Response,
-    resolved: ResolvedAudit | undefined,
-    account: { id: number },
-    // Only the id and name are recorded, so a caller that has just those - a conditional audit reading them
-    // off the request - can emit without a full DBEnvironment.
-    environment: { id: number; name: string } | undefined,
-    // Supplied when the request cannot identify the caller but the handler can - an OAuth callback recovers
-    // its end user from the connect session it looked up.
-    actorOverride?: AuditActor
-): Promise<void> {
-    // Stamp occurredAt now so it reflects the response time, not audit-write latency.
-    const occurredAt = new Date().toISOString();
-    try {
-        const locals = res.locals as Partial<RequestLocals>;
-        const target = resolved?.target;
-        const metadata = resolved?.metadata;
-        const event = {
-            occurredAt,
-            accountId: account.id,
-            environment: policy.scope === 'account' || !environment ? null : { id: environment.id, display: environment.name },
-            actor: actorOverride ?? resolveActor(locals),
-            resource: policy.resource,
-            action: policy.action,
-            targets: Array.isArray(target) ? target : target ? [target] : [],
-            ...auditRequestFields(req, account.id),
-            outcome: outcomeFromStatus(res.statusCode),
-            ...(metadata ? { metadata } : {})
-        } as AuditEvent;
-        await recordAuditEvent(event);
-    } catch (err) {
-        logger.error(`failed to emit audit event`, err);
-        auditEventDropped(policy.resource, 'build_failed');
-    }
-}
-
-export function resolveAuditAttribution(req: Request, locals: Partial<RequestLocals>): AuditAttribution {
-    return { kind: 'request', actor: resolveActor(locals), context: contextFromRequest(req) };
-}
-
-// The entitlement belongs to the audited account, which a spec may override. Reuse the caller's plan
-// when they are the same account, so only an override pays for a lookup.
-async function auditedAccountPlan(account: { id: number }, locals: Partial<RequestLocals>) {
-    return account.id === locals.account?.id ? locals.plan : await getPlanSafe(db.knex, { accountId: account.id });
-}
-
-interface AuditSubject {
-    account: { id: number; uuid: string };
-    environment: { id: number; name: string } | undefined;
-}
-
-interface ResolvedAudit {
-    target: AuditTarget | AuditTarget[] | undefined;
-    metadata: unknown;
-}
-
-// Place AFTER auth and BEFORE authorization so it captures every outcome — including 403 denials
-// that never reach the controller.
-export function auditable<TEndpoint extends AuditableEndpoint>(spec: AuditSpec<TEndpoint>): RequestHandler {
-    return build(spec);
-}
-
-/**
- * For an endpoint that only sometimes does the audited thing — a connection upsert that may have
- * re-authorized rather than created. `skipWhen` is read at finish, so it sees what the handler recorded on
- * the request, and it is deliberately a negative condition: the default stays "emit", so a denial or a
- * failure is still recorded and only the named case is dropped.
- *
- * A separate factory rather than a field on `AuditSpec`, so `auditable`'s promise to record every outcome
- * cannot be weakened one spec at a time.
- */
-export function maybeAuditable<TEndpoint extends AuditableEndpoint>(
-    spec: AuditSpec<TEndpoint> & {
-        skipWhen: (req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => boolean;
-        subject: (req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => AuditSubject | undefined;
-        actor?: ((req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => AuditActor) | undefined;
-        // `target`/`metadata` are resolved before the handler runs and `*FromResponse` needs a JSON body, so
-        // neither can read what the handler recorded — and the OAuth callback answers with HTML. This fills
-        // whatever is still missing, at finish, from the request.
-        atFinish?: (req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => ResolvedAudit;
-    }
-): RequestHandler {
-    const { skipWhen, atFinish, subject, actor, ...rest } = spec;
-    return build(rest as AuditSpec<TEndpoint>, { skipWhen, atFinish, subject, actor });
-}
-
-function build<TEndpoint extends AuditableEndpoint>(
-    spec: AuditSpec<TEndpoint>,
-    conditional?: {
-        skipWhen: (req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => boolean;
-        atFinish?: ((req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => ResolvedAudit) | undefined;
-        // Who the event belongs to, read at finish. `auditable` takes this from `res.locals` before the
-        // handler runs, which an unauthenticated route cannot supply.
-        subject: (req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => AuditSubject | undefined;
-        actor?: ((req: AuditRequest<TEndpoint>, locals: Partial<RequestLocals>) => AuditActor) | undefined;
-    }
-): RequestHandler {
-    return (req, res, next) => {
-        void (async () => {
-            try {
-                const locals = res.locals as Partial<RequestLocals>;
-                if (conditional) {
-                    // Everything this decides needs the handler to have run: whether the audited thing
-                    // happened at all, and — on an unauthenticated route — which account it happened to. So
-                    // the listener goes on unconditionally and the entitlement is checked at finish.
-                    res.on('finish', () => {
-                        void (async () => {
-                            const typedReq = req as AuditRequest<TEndpoint>;
-                            if (conditional.skipWhen(typedReq, locals)) {
-                                return;
-                            }
-                            const subject = conditional.subject(typedReq, locals);
-                            if (!subject || !(await canRecordAuditTrail(subject.account.uuid, await auditedAccountPlan(subject.account, locals)))) {
-                                return;
-                            }
-                            await emit(
-                                spec.policy,
-                                req,
-                                res,
-                                conditional.atFinish?.(typedReq, locals),
-                                subject.account,
-                                subject.environment,
-                                conditional.actor?.(typedReq, locals)
-                            );
-                        })();
-                    });
-                    return;
-                }
-                // Resolve the audited account before the flag gate: a spec may attribute the event to an
-                // account other than the caller's (see AuditSpec.account), and the gate must use that one.
-                const account = spec.account ? await spec.account(req, locals) : locals.account;
-                // Freeze account + environment before the handler runs, for the same reason as target/metadata below.
-                const environment = spec.environment ? await resolveEnvironment(spec.environment, req, locals, spec.policy.resource) : locals.environment;
-                if (account && (await canRecordAuditTrail(account.uuid, await auditedAccountPlan(account, locals)))) {
-                    // Capture the response body only when a spec needs it — the id of a created resource is
-                    // known only after the handler responds. Wrap res.json before next() runs the handler.
-                    let responseBody: unknown;
-                    if (spec.targetFromResponse || spec.metadataFromResponse) {
-                        const originalJson = res.json.bind(res);
-                        res.json = ((body: unknown) => {
-                            responseBody = body;
-                            return originalJson(body);
-                        }) as typeof res.json;
-                    }
-                    // Register the finish listener only once we know we should audit — a disabled account
-                    // never installs a dead listener. It reads `resolved` lazily at finish, so it captures
-                    // whatever each resolver managed to produce.
-                    let resolved: ResolvedAudit | undefined;
-                    res.on('finish', () => {
-                        void (async () => {
-                            if (outcomeFromStatus(res.statusCode) === 'success' && responseBody !== undefined && resolved) {
-                                if (spec.targetFromResponse && resolved.target === undefined) {
-                                    try {
-                                        resolved.target = await spec.targetFromResponse(responseBody as TEndpoint['Success'], req, locals);
-                                    } catch (err) {
-                                        auditEnrichmentFailed('target', spec.policy.resource, err);
-                                    }
-                                }
-                                if (spec.metadataFromResponse) {
-                                    try {
-                                        const fromResponse = await spec.metadataFromResponse(responseBody as TEndpoint['Success'], req, locals);
-                                        resolved.metadata = omitUndefined({
-                                            ...(resolved.metadata && typeof resolved.metadata === 'object' ? resolved.metadata : {}),
-                                            ...fromResponse
-                                        });
-                                    } catch (err) {
-                                        auditEnrichmentFailed('metadata', spec.policy.resource, err);
-                                    }
-                                }
-                            }
-                            await emit(spec.policy, req, res, resolved, account, environment);
-                        })();
-                    });
-                    // Resolve target and metadata before the handler runs — some handlers move or overwrite
-                    // the pre-mutation state (a removed member, an old role).
-                    const partial: ResolvedAudit = { target: undefined, metadata: undefined };
-                    try {
-                        partial.target = spec.target ? await spec.target(req, locals) : undefined;
-                    } catch (err) {
-                        auditEnrichmentFailed('target', spec.policy.resource, err);
-                    }
-                    try {
-                        partial.metadata = spec.metadata ? await spec.metadata(req, locals) : undefined;
-                    } catch (err) {
-                        auditEnrichmentFailed('metadata', spec.policy.resource, err);
-                    }
-                    resolved = partial;
-                }
-            } catch (err) {
-                logger.error(`failed to build audit event`, err);
-            } finally {
-                next();
-            }
-        })();
-    };
-}
-
-function param(req: Request<any, any, any, any>, key: string): unknown {
-    return (req.params as Record<string, unknown>)[key];
-}
-function query(req: Request<any, any, any, any>, key: string): unknown {
-    return (req.query as Record<string, unknown>)[key];
-}
-function bodyField(req: Request<any, any, any, any>, key: string): unknown {
-    return (req.body as Record<string, unknown> | undefined)?.[key];
-}
-/** Resolvers run before the controller validates, so a field the endpoint types as a string can hold anything. */
-function positiveInt(value: unknown): number | undefined {
-    const parsed = typeof value === 'number' || (typeof value === 'string' && value.trim() !== '') ? Number(value) : NaN;
-    return Number.isInteger(parsed) && parsed > 0 ? parsed : undefined;
-}
-function nonEmptyString(value: unknown): string | undefined {
-    return typeof value === 'string' && value.length > 0 ? value : undefined;
-}
-function providerConfigKeyMeta(value: unknown): Record<string, unknown> | undefined {
-    return omitUndefined({ providerConfigKey: nonEmptyString(value) });
-}
-function connectionUpdatedMeta(providerConfigKey: string | undefined, fields: string[] | undefined): Record<string, unknown> | undefined {
-    return omitUndefined({
-        providerConfigKey: providerConfigKey && providerConfigKey.length > 0 ? providerConfigKey : undefined,
-        changedFields: fields
-    });
-}
-function syncFrequencyMeta(frequency: unknown): Record<string, unknown> | undefined {
-    return omitUndefined({ frequency: nonEmptyString(frequency) });
-}
-function functionDeletedMeta(type: unknown): Record<string, unknown> | undefined {
-    // A sync and an action can share a name; `type` disambiguates which function was deleted.
-    return omitUndefined({ type: nonEmptyString(type) });
-}
-// Keep only the origin (scheme + host) of a URL — a webhook URL can carry a secret token in its path,
-// query string, or userinfo, and this goes into the immutable audit record.
-function safeUrl(value: unknown): string | undefined {
-    if (typeof value !== 'string' || value.length === 0) {
-        return undefined;
-    }
-    try {
-        const url = new URL(value);
-        return url.origin;
-    } catch {
-        return undefined;
-    }
-}
-// Target whose display is looked up from the DB best-effort; failures degrade to no display.
-async function dbTarget(type: AuditTargetType, value: unknown, lookup: (id: string) => Promise<string | undefined>): Promise<AuditTarget | undefined> {
-    const id = toId(value);
-    if (!id) {
-        return undefined;
-    }
-    const display = await resolveDisplay(type, () => lookup(id));
-    return { type, id, ...(display ? { display } : {}) };
-}
-
-function memberTarget(req: Request<{ id: number }>, locals: Partial<RequestLocals>): Promise<AuditTarget | undefined> {
-    return dbTarget('member', req.params.id, async (id) => {
-        if (!locals.account) {
-            return undefined;
-        }
-        const user = await userService.getUserByIdAndAccountId(Number(id), locals.account.id);
-        return user?.email;
-    });
-}
-
-async function environmentFromBody(value: unknown, locals: Partial<RequestLocals>): Promise<{ id: number; name: string } | null> {
-    const environmentId = positiveInt(value);
-    if (environmentId === undefined || !locals.account) {
-        return null;
-    }
-    const environment = await environmentService.getByIdWithoutSecrets(environmentId, locals.account.id);
-    return environment ? { id: environment.id, name: environment.name } : null;
-}
-function integrationTarget(value: unknown, locals: Partial<RequestLocals>): Promise<AuditTarget | undefined> {
-    return dbTarget('integration', value, async (id) => {
-        if (!locals.environment) {
-            return undefined;
-        }
-        const summary = await configService.getIntegrationSummary(locals.environment.id, id);
-        return summary?.display_name ?? undefined;
-    });
-}
-async function integrationProviderMeta(value: unknown, locals: Partial<RequestLocals>): Promise<Record<string, unknown> | undefined> {
-    const key = nonEmptyString(value);
-    if (!key || !locals.environment) {
-        return undefined;
-    }
-    try {
-        const summary = await configService.getIntegrationSummary(locals.environment.id, key);
-        return omitUndefined({ provider: summary?.provider });
-    } catch (err) {
-        auditEnrichmentFailed('metadata', 'integration', err);
-        return undefined;
-    }
-}
-function apiKeyTarget(value: unknown, locals: Partial<RequestLocals>): Promise<AuditTarget | undefined> {
-    return dbTarget('api_key', value, async (id) => {
-        if (!locals.environment) {
-            return undefined;
-        }
-        const result = await customerKeyService.getApiKeysByEnv(db.knex, locals.environment.id);
-        if (result.isErr()) {
-            throw result.error;
-        }
-        return result.value.find((key) => String(key.id) === id)?.display_name;
-    });
-}
-
-function accountApiKeyTarget(value: unknown, locals: Partial<RequestLocals>): Promise<AuditTarget | undefined> {
-    return dbTarget('api_key', value, async (id) => {
-        const numericId = Number(id);
-        // Audit runs before controller param validation; skip the DB lookup for non-numeric
-        // keyIds so malformed deletes return 400 without an audit display-resolution warning.
-        if (Number.isNaN(numericId) || !locals.account) {
-            return undefined;
-        }
-        const result = await customerKeyService.getAccountApiKeyDisplayName(db.knex, numericId, locals.account.id);
-        if (result.isErr()) {
-            throw result.error;
-        }
-        return result.value;
-    });
-}
-
-function publicEnvApiKeyTarget(keyId: unknown, environmentId: unknown, locals: Partial<RequestLocals>): Promise<AuditTarget | undefined> {
-    return dbTarget('api_key', keyId, async (id) => {
-        const numericId = positiveInt(id);
-        const numericEnvId = positiveInt(environmentId);
-        if (numericId === undefined || numericEnvId === undefined || !locals.account) {
-            return undefined;
-        }
-        const result = await customerKeyService.getApiKeyDisplayName(db.knex, numericId, numericEnvId, locals.account.id);
-        if (result.isErr()) {
-            throw result.error;
-        }
-        return result.value;
-    });
-}
-
-function accountEnvironmentTarget(value: unknown, locals: Partial<RequestLocals>): Promise<AuditTarget | undefined> {
-    return dbTarget('environment', value, async (id) => {
-        const numericId = Number(id);
-        if (Number.isNaN(numericId) || !locals.account) {
-            return undefined;
-        }
-        const environment = await environmentService.getByIdWithoutSecrets(numericId, locals.account.id);
-        return environment?.name;
-    });
-}
-
-export const auditConnectionRefreshed = auditable<PostConnectionRefresh>({
-    policy: Audit.auditable({ resource: 'connection', action: 'refreshed', scope: 'environment' }),
-    target: (req) => makeTarget('connection', req.params.connectionId),
-    metadata: (req) => providerConfigKeyMeta(req.query.provider_config_key)
-});
-export const auditConnectionUpdated = auditable<PatchConnection>({
-    policy: Audit.auditable({ resource: 'connection', action: 'updated', scope: 'environment' }),
-    target: (req) => makeTarget('connection', req.params.connectionId),
-    metadata: (req) => connectionUpdatedMeta(req.query.provider_config_key, changedFields(req.body))
-});
-// Parsed with the endpoint's own schema, so the filters recorded are the ones it accepted rather than a
-// second, looser reading of the same query.
-function auditTrailFilters(data: { from?: string | undefined; to?: string | undefined; resources?: string[] | undefined; actions?: string[] | undefined }) {
-    return { from: data.from, to: data.to, resources: data.resources, actions: data.actions };
-}
-
-export const auditTrailExported = auditable<GetAuditTrailExport>({
-    policy: Audit.auditable({ resource: 'audit_trail', action: 'exported', scope: 'account' }),
-    metadata: (req) => {
-        const query = auditExportQuery.safeParse(req.query);
-        return query.success ? auditTrailFilters(query.data) : undefined;
-    }
-});
-
-export const auditTrailQueried = auditable<GetAuditTrail>({
-    policy: Audit.auditable({ resource: 'audit_trail', action: 'queried', scope: 'account' }),
-    // `continued` marks a page of an earlier query, so an auditor can collapse one browsing session. It is not
-    // a gate: a cursor is unsigned, so letting it decide whether we record would let a reader opt out.
-    metadata: (req) => {
-        const query = auditListQuery.safeParse(req.query);
-        return query.success ? { ...auditTrailFilters(query.data), ...(query.data.cursor ? { continued: true } : {}) } : undefined;
-    }
-});
-
-/**
- * `connection.created` for every route that upserts a connection. One middleware rather than fifteen
- * near-identical ones, because the handler reports what happened on the request and nothing here is
- * endpoint-specific. The trade-off is that the policy is pinned to this shape instead of being checked
- * against each route's declared `Audit`, which a shared middleware cannot do.
- *
- * A re-authorization is skipped: it upserts through the same endpoints and answers the same 200.
- * `connection.reauthorized` exists in the vocabulary if we ever want it, and it belongs here.
- */
-export const auditConnectionCreated = maybeAuditable<Endpoint<any> & { Audit: AuditPolicy<'connection', 'created', 'environment'> }>({
-    policy: Audit.auditable({ resource: 'connection', action: 'created', scope: 'environment' }),
-    skipWhen: (req) => req.audit?.connectionUpsert?.operation === 'override',
-    subject: (req, locals) => {
-        const upsert = req.audit?.connectionUpsert;
-        if (upsert) {
-            return { account: upsert.account, environment: upsert.environment };
-        }
-        // Nothing was created, so this is a denial or a failure, and the account comes from the request
-        // instead. An attempt on an unauthenticated route has neither, and one nobody can be attributed to
-        // is not worth a row.
-        return locals.account ? { account: locals.account, environment: locals.environment } : undefined;
-    },
-    // The request wins when it proves who called; a connect session's end user reaches an unauthenticated
-    // callback only through the handler, so it fills the gap the request leaves.
-    actor: (req, locals) => connectionCreatedActor(resolveActor(locals), req.audit?.connectionUpsert?.endUser),
-    atFinish: (req) => {
-        const upsert = req.audit?.connectionUpsert;
-        const connectionId = upsert?.connectionId ?? nonEmptyString(query(req, 'connection_id')) ?? nonEmptyString(bodyField(req, 'connection_id'));
-        const providerConfigKey =
-            upsert?.providerConfigKey ?? nonEmptyString(param(req, 'providerConfigKey')) ?? nonEmptyString(bodyField(req, 'provider_config_key'));
-        return {
-            target: makeTarget('connection', connectionId),
-            metadata: providerConfigKeyMeta(providerConfigKey)
-        };
-    }
-});
-
-export const auditPublicConnectionUpdated = auditable<PatchPublicConnection>({
-    policy: Audit.auditable({ resource: 'connection', action: 'updated', scope: 'environment' }),
-    target: (req) => makeTarget('connection', req.params.connectionId),
-    metadata: (req) => connectionUpdatedMeta(req.query.provider_config_key, changedFields(req.body))
-});
-export const auditConnectionMetadataUpdated = auditable<PostConnectionMetadata>({
-    policy: Audit.auditable({ resource: 'connection', action: 'metadata_updated', scope: 'environment' }),
-    target: (req) => makeTarget('connection', req.params.connectionId),
-    metadata: (req) => providerConfigKeyMeta(req.query.provider_config_key)
-});
-export const auditConnectionDeleted = auditable<DeleteConnection>({
-    policy: Audit.auditable({ resource: 'connection', action: 'deleted', scope: 'environment' }),
-    target: (req) => makeTarget('connection', req.params.connectionId),
-    metadata: (req) => providerConfigKeyMeta(req.query.provider_config_key)
-});
-export const auditPublicConnectionDeleted = auditable<DeletePublicConnection>({
-    policy: Audit.auditable({ resource: 'connection', action: 'deleted', scope: 'environment' }),
-    target: (req) => makeTarget('connection', req.params.connectionId),
-    metadata: (req) => providerConfigKeyMeta(req.query.provider_config_key)
-});
-
-export const auditIntegrationUpdated = auditable<PatchIntegration>({
-    policy: Audit.auditable({ resource: 'integration', action: 'updated', scope: 'environment' }),
-    target: (req, locals) => integrationTarget(req.params.providerConfigKey, locals),
-    metadata: async (req, locals) =>
-        omitUndefined({ ...(await integrationProviderMeta(req.params.providerConfigKey, locals)), changedFields: changedFields(req.body) })
-});
-export const auditPublicIntegrationUpdated = auditable<PatchPublicIntegration>({
-    policy: Audit.auditable({ resource: 'integration', action: 'updated', scope: 'environment' }),
-    target: (req, locals) => integrationTarget(req.params.uniqueKey, locals),
-    metadata: async (req, locals) => omitUndefined({ ...(await integrationProviderMeta(req.params.uniqueKey, locals)), changedFields: changedFields(req.body) })
-});
-export const auditIntegrationDeleted = auditable<DeleteIntegration>({
-    policy: Audit.auditable({ resource: 'integration', action: 'deleted', scope: 'environment' }),
-    target: (req, locals) => integrationTarget(req.params.providerConfigKey, locals),
-    metadata: (req, locals) => integrationProviderMeta(req.params.providerConfigKey, locals)
-});
-export const auditPublicIntegrationDeleted = auditable<DeletePublicIntegration>({
-    policy: Audit.auditable({ resource: 'integration', action: 'deleted', scope: 'environment' }),
-    target: (req, locals) => integrationTarget(req.params.uniqueKey, locals),
-    metadata: (req, locals) => integrationProviderMeta(req.params.uniqueKey, locals)
-});
-
-export const auditFunctionDeleted = auditable<DeleteIntegrationFunction>({
-    policy: Audit.auditable({ resource: 'function', action: 'deleted', scope: 'environment' }),
-    target: (req) => makeTarget('function', functionTargetId(req.params.providerConfigKey, req.params.functionName)),
-    metadata: (req) => functionDeletedMeta(req.query.type)
-});
-export const auditPublicFunctionDeleted = auditable<DeletePublicIntegrationFunction>({
-    policy: Audit.auditable({ resource: 'function', action: 'deleted', scope: 'environment' }),
-    target: (req) => makeTarget('function', functionTargetId(req.params.uniqueKey, req.params.name)),
-    metadata: (req) => functionDeletedMeta(req.query.type)
-});
-
-export const auditApiKeyUpdated = auditable<PatchApiKey>({
-    policy: Audit.auditable({ resource: 'api_key', action: 'updated', scope: 'environment' }),
-    target: (req, locals) => apiKeyTarget(req.params.keyId, locals),
-    metadata: (req) =>
-        omitUndefined({
-            displayName: typeof req.body.display_name === 'string' ? req.body.display_name : undefined,
-            scopes: Array.isArray(req.body.scopes) ? req.body.scopes.filter((s) => typeof s === 'string') : undefined
-        })
-});
-export const auditApiKeyDeleted = auditable<DeleteApiKey>({
-    policy: Audit.auditable({ resource: 'api_key', action: 'deleted', scope: 'environment' }),
-    target: (req, locals) => apiKeyTarget(req.params.keyId, locals)
-});
-export const auditPublicApiKeyDeleted = auditable<DeletePublicApiKey>({
-    policy: Audit.auditable({ resource: 'api_key', action: 'deleted', scope: 'environment' }),
-    environment: (req, locals) => environmentFromBody(req.body.environment_id, locals),
-    target: (req, locals) => publicEnvApiKeyTarget(req.body.key_id, req.body.environment_id, locals)
-});
-export const auditAccountApiKeyDeleted = auditable<DeleteAccountApiKey>({
-    policy: Audit.auditable({ resource: 'api_key', action: 'deleted', scope: 'account' }),
-    target: (req, locals) => accountApiKeyTarget(req.params.keyId, locals)
-});
-
-export const auditSyncEnabled = auditable<PatchFlowEnable>({
-    policy: Audit.auditable({ resource: 'sync', action: 'enabled', scope: 'environment' }),
-    target: (req) => makeTarget('sync', req.body.scriptName),
-    metadata: (req) => syncBaseMeta(req.body.providerConfigKey)
-});
-export const auditSyncDisabled = auditable<PatchFlowDisable>({
-    policy: Audit.auditable({ resource: 'sync', action: 'disabled', scope: 'environment' }),
-    target: (req) => makeTarget('sync', req.body.scriptName),
-    metadata: (req) => syncBaseMeta(req.body.providerConfigKey)
-});
-export const auditSyncFrequencyChanged = auditable<PatchFlowFrequency>({
-    policy: Audit.auditable({ resource: 'sync', action: 'frequency_changed', scope: 'environment' }),
-    target: (req) => makeTarget('sync', req.body.scriptName),
-    metadata: (req) => ({ ...syncBaseMeta(req.body.providerConfigKey), ...syncFrequencyMeta(req.body.frequency) })
-});
-export const auditPublicSyncFrequencyChanged = auditable<PutPublicSyncConnectionFrequency>({
-    policy: Audit.auditable({ resource: 'sync', action: 'frequency_changed', scope: 'environment' }),
-    target: (req) => makeTarget('sync', syncTargetId(req.body.sync_name, req.body.sync_variant)),
-    metadata: (req) => ({
-        ...syncBaseMeta(req.body.provider_config_key, req.body.connection_id),
-        ...syncFrequencyMeta(req.body.frequency)
-    })
-});
-export const auditSyncVariantCreated = auditable<PostSyncVariant>({
-    policy: Audit.auditable({ resource: 'sync', action: 'variant_created', scope: 'environment' }),
-    target: (req) => makeTarget('sync', syncTargetId(req.params.name, req.params.variant)),
-    metadata: (req) => ({ variant: req.params.variant, ...syncBaseMeta(req.body.provider_config_key, req.body.connection_id) })
-});
-export const auditSyncVariantDeleted = auditable<DeleteSyncVariant>({
-    policy: Audit.auditable({ resource: 'sync', action: 'variant_deleted', scope: 'environment' }),
-    target: (req) => makeTarget('sync', syncTargetId(req.params.name, req.params.variant)),
-    metadata: (req) => ({ variant: req.params.variant, ...syncBaseMeta(req.body.provider_config_key, req.body.connection_id) })
-});
-
-export const auditMemberRemoved = auditable<DeleteTeamUser>({
-    policy: Audit.auditable({ resource: 'member', action: 'removed', scope: 'account' }),
-    target: memberTarget
-});
-export const auditMemberRoleChanged = auditable<PatchTeamUser>({
-    policy: Audit.auditable({ resource: 'member', action: 'role_changed', scope: 'account' }),
-    target: memberTarget,
-    metadata: async (req, locals) => {
-        const role = req.body.role;
-        let fromRole: string | undefined;
-        const id = toId(req.params.id);
-        if (id && locals.account) {
-            const accountId = locals.account.id;
-            fromRole = await resolveDisplay('member', async () => {
-                const user = await userService.getUserByIdAndAccountId(Number(id), accountId);
-                return user?.role;
-            });
-        }
-        return omitUndefined({
-            toRole: typeof role === 'string' ? role : undefined,
-            fromRole: fromRole ? fromRole : undefined
-        });
-    }
-});
-export const auditTeamUpdated = auditable<PutTeam>({
-    policy: Audit.auditable({ resource: 'team', action: 'updated', scope: 'account' }),
-    target: (_req, locals) => makeTarget('team', locals.account?.id, locals.account?.name),
-    metadata: (req) => {
-        const name = req.body.name;
-        return typeof name === 'string' ? { name } : undefined;
-    }
-});
-export const auditUserUpdated = auditable<PatchUser>({
-    policy: Audit.auditable({ resource: 'user', action: 'updated', scope: 'account' }),
-    target: (_req, locals) => makeTarget('user', locals.user?.id, locals.user?.email),
-    metadata: (req) =>
-        omitUndefined({
-            name: nonEmptyString(req.body.name),
-            gettingStartedClosed: typeof req.body.gettingStartedClosed === 'boolean' ? req.body.gettingStartedClosed : undefined
-        })
-});
-
-export const auditEnvironmentDeleted = auditable<DeleteEnvironment>({
-    policy: Audit.auditable({ resource: 'environment', action: 'deleted', scope: 'environment' }),
-    target: (_req, locals) => makeTarget('environment', locals.environment?.id, locals.environment?.name)
-});
-export const auditPublicEnvironmentDeleted = auditable<DeletePublicEnvironment>({
-    policy: Audit.auditable({ resource: 'environment', action: 'deleted', scope: 'account' }),
-    target: (req, locals) => accountEnvironmentTarget(req.params.environmentId, locals)
-});
-export const auditEnvironmentUpdated = auditable<PatchEnvironment>({
-    policy: Audit.auditable({ resource: 'environment', action: 'updated', scope: 'environment' }),
-    target: (_req, locals) => makeTarget('environment', locals.environment?.id, locals.environment?.name),
-    metadata: (req) =>
-        omitUndefined({
-            name: nonEmptyString(req.body.name),
-            changedFields: changedFields(req.body)
-        })
-});
-export const auditEnvironmentVariablesChanged = auditable<PostEnvironmentVariables>({
-    policy: Audit.auditable({ resource: 'environment', action: 'variables_changed', scope: 'environment' }),
-    target: (_req, locals) => makeTarget('environment', locals.environment?.id, locals.environment?.name),
-    metadata: (req) => {
-        const variables = req.body.variables;
-        if (!Array.isArray(variables)) {
-            return undefined;
-        }
-        const variableNames = variables
-            .map((v) => (v && typeof v === 'object' ? (v as Record<string, unknown>)['name'] : undefined))
-            .filter((n): n is string => typeof n === 'string');
-        return omitUndefined({ variableCount: variables.length, variableNames: variableNames.length > 0 ? variableNames : undefined });
-    }
-});
-export const auditWebhookSigningKeyRotated = auditable<PostRotateWebhookSigningKey>({
-    policy: Audit.auditable({ resource: 'environment', action: 'webhook_signing_key_rotated', scope: 'environment' }),
-    target: (_req, locals) => makeTarget('environment', locals.environment?.id, locals.environment?.name)
-});
-export const auditPublicWebhookSigningKeyRotated = auditable<PostPublicRotateWebhookSigningKey>({
-    policy: Audit.auditable({ resource: 'environment', action: 'webhook_signing_key_rotated', scope: 'environment' }),
-    target: (_req, locals) => makeTarget('environment', locals.environment?.id, locals.environment?.name)
-});
-export const auditEnvironmentWebhookUrlsChanged = auditable<PatchWebhook>({
-    policy: Audit.auditable({ resource: 'environment', action: 'webhook_urls_changed', scope: 'environment' }),
-    target: (_req, locals) => makeTarget('environment', locals.environment?.id, locals.environment?.name),
-    metadata: (req) =>
-        omitUndefined({
-            changedFields: changedFields(req.body),
-            primaryUrl: safeUrl(req.body.primary_url),
-            secondaryUrl: safeUrl(req.body.secondary_url)
-        })
-});
-
-export const auditBillingPlanChanged = auditable<PostPlanChange>({
-    policy: Audit.auditable({ resource: 'billing', action: 'plan_changed', scope: 'account' }),
-    metadata: (req, locals) =>
-        omitUndefined({
-            toPlan: typeof req.body.orbId === 'string' ? req.body.orbId : undefined,
-            fromPlan: locals.plan?.name || undefined
-        })
-});
-export const auditBillingTrialExtended = auditable<PostPlanExtendTrial>({
-    policy: Audit.auditable({ resource: 'billing', action: 'trial_extended', scope: 'account' })
-});
-export const auditBillingDetailsChanged = auditable<PutBillingInvoicingDetails>({
-    policy: Audit.auditable({ resource: 'billing', action: 'details_changed', scope: 'account' })
-});
-export const auditBillingSpendAlertChanged = auditable<PutSpendAlert>({
-    policy: Audit.auditable({ resource: 'billing', action: 'spend_alert_changed', scope: 'account' }),
-    metadata: (req) => omitUndefined({ thresholdInCents: typeof req.body.thresholdInCents === 'number' ? req.body.thresholdInCents : undefined })
-});
-export const auditBillingSpendAlertRemoved = auditable<DeleteSpendAlert>({
-    policy: Audit.auditable({ resource: 'billing', action: 'spend_alert_removed', scope: 'account' })
-});
-// SetupIntent only — pm id isn't known yet (arrives via webhook); response is just a client secret, so nothing to record.
-export const auditBillingPaymentMethodAdded = auditable<PostStripeCollectPayment>({
-    policy: Audit.auditable({ resource: 'billing', action: 'payment_method_added', scope: 'account' })
-});
-export const auditBillingPaymentMethodRemoved = auditable<DeleteStripePayment>({
-    policy: Audit.auditable({ resource: 'billing', action: 'payment_method_removed', scope: 'account' }),
-    metadata: (req) =>
-        typeof req.query.payment_id === 'string' && req.query.payment_id.length > 0 && req.query.payment_id.length <= 255
-            ? { paymentMethodId: req.query.payment_id }
-            : undefined
-});
-
-export const auditAppAuthPasswordChanged = auditable<PutUserPassword>({
-    policy: Audit.auditable({ resource: 'app_auth', action: 'password_changed', scope: 'account' }),
-    target: (_req, locals) => makeTarget('user', locals.user?.id, locals.user?.email)
-});
-
-function functionTargetId(integrationId: unknown, name: unknown): string | undefined {
-    const integration = nonEmptyString(integrationId);
-    const functionName = nonEmptyString(name);
-    return integration && functionName ? `${integration}:${functionName}` : functionName;
-}
-
-/** `base` is the default variant, so it is left out of the id rather than spelled out. */
-export function syncTargetId(name: string, variant?: string): string {
-    return variant && variant !== 'base' ? `${name}::${variant}` : name;
-}
-export function syncBaseMeta(providerConfigKey: unknown, connectionId?: unknown): Record<string, unknown> | undefined {
-    return omitUndefined({ providerConfigKey: nonEmptyString(providerConfigKey), connectionId: nonEmptyString(connectionId) });
-}
-function syncTargetsFromBody(syncs: (string | { name: string; variant: string })[] | undefined): AuditTarget[] | undefined {
-    if (!Array.isArray(syncs)) {
-        return undefined;
-    }
-    const targets = normalizeSyncParams(syncs)
-        .map(({ syncName, syncVariant }) => makeTarget('sync', syncTargetId(syncName, syncVariant)))
-        .filter((t): t is AuditTarget => Boolean(t));
-    return targets.length > 0 ? targets : undefined;
-}
-
-export const auditIntegrationCreated = auditable<PostIntegration>({
-    policy: Audit.auditable({ resource: 'integration', action: 'created', scope: 'environment' }),
-    // The final unique_key is only certain in the response — the private path omits it from the request.
-    targetFromResponse: (response) => makeTarget('integration', response.data.unique_key, response.data.display_name ?? undefined),
-    metadata: (req) => omitUndefined({ provider: nonEmptyString(req.body.provider) })
-});
-export const auditPublicIntegrationCreated = auditable<PostPublicIntegration>({
-    policy: Audit.auditable({ resource: 'integration', action: 'created', scope: 'environment' }),
-    targetFromResponse: (response) => makeTarget('integration', response.data.unique_key, response.data.display_name ?? undefined),
-    metadata: (req) => omitUndefined({ provider: nonEmptyString(req.body.provider) })
-});
-export const auditPublicQuickstartIntegrationCreated = auditable<PostPublicQuickstartIntegration>({
-    policy: Audit.auditable({ resource: 'integration', action: 'created', scope: 'environment' }),
-    targetFromResponse: (response) => makeTarget('integration', response.data.unique_key, response.data.display_name ?? undefined),
-    metadata: (req) => omitUndefined({ provider: nonEmptyString(req.body.provider) })
-});
-
-export const auditEnvironmentCreated = auditable<PostEnvironment>({
-    policy: Audit.auditable({ resource: 'environment', action: 'created', scope: 'account' }),
-    targetFromResponse: (response) => makeTarget('environment', response.data.id, response.data.name),
-    metadata: (req) => omitUndefined({ name: req.body.name })
-});
-export const auditPublicEnvironmentCreated = auditable<PostPublicEnvironment>({
-    policy: Audit.auditable({ resource: 'environment', action: 'created', scope: 'account' }),
-    targetFromResponse: (response) => makeTarget('environment', response.data.id, response.data.name),
-    metadata: (req) => omitUndefined({ name: req.body.name })
-});
-
-export const auditApiKeyCreated = auditable<CreateApiKey>({
-    policy: Audit.auditable({ resource: 'api_key', action: 'created', scope: 'environment' }),
-    // Never read the secret from the response — only the id and display name identify the key.
-    targetFromResponse: (response) => makeTarget('api_key', response.data.id, response.data.display_name),
-    metadata: (req) => omitUndefined({ displayName: req.body.display_name }),
-    metadataFromResponse: (response) => omitUndefined({ scopes: response.data.scopes })
-});
-export const auditPublicApiKeyCreated = auditable<PostPublicApiKey>({
-    policy: Audit.auditable({ resource: 'api_key', action: 'created', scope: 'environment' }),
-    environment: (req, locals) => environmentFromBody(req.body.environment_id, locals),
-    targetFromResponse: (response) => makeTarget('api_key', response.data.id, response.data.display_name),
-    metadata: (req) => omitUndefined({ displayName: nonEmptyString(req.body.display_name) }),
-    metadataFromResponse: (response) => omitUndefined({ scopes: response.data.scopes })
-});
-export const auditAccountApiKeyCreated = auditable<CreateAccountApiKey>({
-    policy: Audit.auditable({ resource: 'api_key', action: 'created', scope: 'account' }),
-    targetFromResponse: (response) => makeTarget('api_key', response.data.id, response.data.display_name),
-    metadata: (req) => omitUndefined({ displayName: req.body.display_name }),
-    // Scopes are chosen by the service/controller today and will be request-configurable later —
-    // always record what was actually persisted.
-    metadataFromResponse: (response) => omitUndefined({ scopes: response.data.scopes })
-});
-
-export const auditMemberInvited = auditable<PostInvite>({
-    policy: Audit.auditable({ resource: 'member', action: 'invited', scope: 'account' }),
-    // An invitee may not have an account at all, so the email is the only identity available.
-    target: (req) =>
-        Array.isArray(req.body.emails)
-            ? req.body.emails.map((email) => makeTarget('member', email, email)).filter((t): t is AuditTarget => Boolean(t))
-            : undefined,
-    metadata: (req) => (req.body.role ? { role: req.body.role } : undefined)
-});
-export const auditMemberInviteRevoked = auditable<DeleteInvite>({
-    policy: Audit.auditable({ resource: 'member', action: 'invite_revoked', scope: 'account' }),
-    target: (req) => makeTarget('member', req.body.email, req.body.email)
-});
-// Recorded under the inviting team (like invited/revoked), not the invitee's own account: locals.account
-// is the accepter's pre-existing account, so resolve the target account from the invitation instead. The
-// invite still exists here (resolved before the handler consumes it).
-async function invitingAccount(req: Request<{ id: string }>): Promise<{ id: number; uuid: string } | undefined> {
-    const invitation = await getInvitation(req.params.id);
-    if (!invitation) {
-        return undefined;
-    }
-    return (await accountService.getAccountById(db.knex, invitation.account_id)) ?? undefined;
-}
-
-// Accept/decline run under webAuth, so the acting user IS the invited member — the actor is resolved
-// from the session and the target email comes from locals, keeping the member identity (email)
-// consistent with the invited/revoked events. The invite token (req.params.id) is not a member identity.
-export const auditMemberInviteAccepted = auditable<AcceptInvite>({
-    policy: Audit.auditable({ resource: 'member', action: 'invite_accepted', scope: 'account' }),
-    account: invitingAccount,
-    target: (_req, locals) => makeTarget('member', locals.user?.id, locals.user?.email)
-});
-export const auditMemberInviteDeclined = auditable<DeclineInvite>({
-    policy: Audit.auditable({ resource: 'member', action: 'invite_declined', scope: 'account' }),
-    account: invitingAccount,
-    target: (_req, locals) => makeTarget('member', locals.user?.id, locals.user?.email)
-});
-
-// A code deploy is performed by the sandbox's CLI and recorded when that reaches /sync/deploy.
-export const auditFunctionDeployedFromTemplate = maybeAuditable<PostFunctionDeployment>({
-    policy: Audit.auditable({ resource: 'function', action: 'deployed', scope: 'environment' }),
-    skipWhen: (req) => req.body.type !== 'template',
-    subject: (_req, locals) => (locals.account ? { account: locals.account, environment: locals.environment } : undefined),
-    atFinish: (req) => ({
-        target: makeTarget('function', functionTargetId(req.body.integration_id, req.body.type === 'template' ? req.body.template : undefined)),
-        metadata: omitUndefined({ source: 'catalog', type: nonEmptyString(req.body.function_type) })
-    })
-});
-export const auditFunctionDeployedCli = auditable<PostDeploy>({
-    policy: Audit.auditable({ resource: 'function', action: 'deployed', scope: 'environment' }),
-    metadata: (req) => omitUndefined({ source: nonEmptyString(req.body.source) ?? 'repo' }),
-    target: (req) =>
-        Array.isArray(req.body.flowConfigs)
-            ? req.body.flowConfigs
-                  .map((flow) => makeTarget('function', functionTargetId(flow.providerConfigKey, flow.syncName)))
-                  .filter((t): t is AuditTarget => Boolean(t))
-            : undefined
-});
-
-function functionBundleTargets(value: unknown): AuditTarget[] | undefined {
-    if (!Array.isArray(value)) {
-        return undefined;
-    }
-
-    return value
-        .map((artifact: unknown) => {
-            if (typeof artifact !== 'object' || artifact === null) {
-                return undefined;
-            }
-            const { integrationId, name } = artifact as { integrationId?: unknown; name?: unknown };
-            if (typeof integrationId !== 'string' || integrationId.length === 0 || typeof name !== 'string' || name.length === 0) {
-                return undefined;
-            }
-            return makeTarget('function', functionTargetId(integrationId, name));
-        })
-        .filter((target): target is AuditTarget => Boolean(target));
-}
-
-export const auditFunctionDeploymentBundle = auditable<PostFunctionDeploymentBundle>({
-    policy: Audit.auditable({ resource: 'function', action: 'deployed', scope: 'environment' }),
-    target: (req) => functionBundleTargets(req.body.functions),
-    metadata: () => ({ type: 'function' })
-});
-
-export const auditPreBuiltDeployed = auditable<PostPreBuiltDeploy>({
-    policy: Audit.auditable({ resource: 'function', action: 'deployed', scope: 'environment' }),
-    target: (req) => makeTarget('function', functionTargetId(req.body.providerConfigKey, req.body.scriptName)),
-    metadata: (req) => omitUndefined({ source: 'catalog', type: nonEmptyString(req.body.type) })
-});
-
-export const auditFunctionUpgraded = auditable<PutUpgradePreBuiltFlow>({
-    policy: Audit.auditable({ resource: 'function', action: 'upgraded', scope: 'environment' }),
-    target: (req) => makeTarget('function', functionTargetId(req.body.providerConfigKey, req.body.scriptName)),
-    metadata: (req) => omitUndefined({ upgradeVersion: nonEmptyString(req.body.upgradeVersion) })
-});
-
-export const auditSyncPaused = auditable<PostPublicSyncPause>({
-    policy: Audit.auditable({ resource: 'sync', action: 'paused', scope: 'environment' }),
-    target: (req) => syncTargetsFromBody(req.body?.syncs),
-    metadata: (req) => syncBaseMeta(req.body.provider_config_key, req.body.connection_id)
-});
-export const auditSyncStarted = auditable<PostPublicSyncStart>({
-    policy: Audit.auditable({ resource: 'sync', action: 'started', scope: 'environment' }),
-    target: (req) => syncTargetsFromBody(req.body?.syncs),
-    metadata: (req) => syncBaseMeta(req.body.provider_config_key, req.body.connection_id)
-});
-
-// MFA factors are per-user and account-scoped; the acting user is always the target. No metadata is
-// recorded — the request bodies carry only TOTP/recovery codes, which must never be persisted.
-export const auditMfaEnrolled = auditable<PostMFAEnrollment>({
-    policy: Audit.auditable({ resource: 'mfa', action: 'enrolled', scope: 'account' }),
-    target: (_req, locals) => makeTarget('user', locals.user?.id, locals.user?.email)
-});
-export const auditMfaEnabled = auditable<PostMFAActivation>({
-    policy: Audit.auditable({ resource: 'mfa', action: 'enabled', scope: 'account' }),
-    target: (_req, locals) => makeTarget('user', locals.user?.id, locals.user?.email)
-});
-export const auditMfaDisabled = auditable<DeleteMFA>({
-    policy: Audit.auditable({ resource: 'mfa', action: 'disabled', scope: 'account' }),
-    target: (_req, locals) => makeTarget('user', locals.user?.id, locals.user?.email)
-});
-export const auditMfaRecoveryRegenerated = auditable<PostMFARecoveryCodes>({
-    policy: Audit.auditable({ resource: 'mfa', action: 'recovery_regenerated', scope: 'account' }),
-    target: (_req, locals) => makeTarget('user', locals.user?.id, locals.user?.email)
-});
-
-const METHOD_BY_TYPE = { code: 'totp', recoveryCode: 'recovery_code' } as const;
-
-// Anchor the event's resource/action to the endpoint's declared Audit policy so this dedicated middleware
-// can't drift from it — the typed auditable() specs get the same guarantee for free via AuditSpec.policy.
-const mfaVerifiedPolicy: PostMFALoginVerification['Audit'] = { kind: 'audit', resource: 'mfa', action: 'verified', scope: 'account' };
-
-// The login-verify route runs BEFORE authentication (the user is mid-login), so res.locals carries no
-// user or account and the standard locals-based auditable() can't attribute the event — and emit()
-// early-returns without an account. Resolve the acting user from the pending-login session (it still
-// exists at middleware entry; the controller deletes it once verification succeeds), load their
-// account directly, and emit an `mfa`/`verified` event on finish for both success and failure.
-export const auditMfaVerified: RequestHandler = (req, res, next) => {
-    // Capture the pending user synchronously — the controller clears the pending session on success.
-    const userId = req.session.pendingMfaLogin?.userId;
-    res.on('finish', () => {
-        void emitMfaVerified(req, res, userId);
-    });
-    next();
-};
-
-async function emitMfaVerified(req: Request, res: Response, pendingUserId: number | undefined): Promise<void> {
-    const occurredAt = new Date().toISOString();
-    try {
-        // Attribute to the pending-login user captured at middleware entry — equal to req.user on success,
-        // and the only attribution available on failure. No pending challenge means there is nothing to audit.
-        if (pendingUserId == null) {
-            return;
-        }
-        const user = await userService.getUserById(pendingUserId, true);
-        if (!user) {
-            return;
-        }
-        const account = await accountService.getAccountById(db.knex, user.account_id);
-        if (!account) {
-            return;
-        }
-        // Runs before authentication, so there is no res.locals.plan to read the entitlement from.
-        if (!(await canRecordAuditTrail(account.uuid, await getPlanSafe(db.knex, { accountId: account.id })))) {
-            return;
-        }
-        const bodyType = (req.body as Partial<PostMFALoginVerification['Body']>)?.type;
-        const method: MfaVerifiedMetadata['method'] | undefined = bodyType && Object.hasOwn(METHOD_BY_TYPE, bodyType) ? METHOD_BY_TYPE[bodyType] : undefined;
-        const event: AuditEvent = {
-            occurredAt,
-            accountId: account.id,
-            environment: null,
-            actor: { type: 'user', id: String(user.id), display: user.email },
-            resource: mfaVerifiedPolicy.resource,
-            action: mfaVerifiedPolicy.action,
-            targets: [{ type: 'user', id: String(user.id), display: user.email }],
-            ...auditRequestFields(req, account.id),
-            outcome: outcomeFromStatus(res.statusCode),
-            ...(method ? { metadata: { method } } : {})
-        };
-        await recordAuditEvent(event);
-    } catch (err) {
-        logger.error(`failed to emit mfa verify audit event`, err);
-        auditEventDropped('mfa', 'build_failed');
-    }
-}
