+++ packages/server/lib/middleware/audit/mfa.middleware.unit.test.ts
@@ -1,0 +1,33 @@
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
+
+import { auditMfaEnabled } from './mfa.middleware.js';
+import { fakeReq, fakeRes, installAuditMockDefaults, locals, resetAuditMocks, runAudit } from './testing.js';
+
+vi.mock('../../audit.js', async (importOriginal) => (await import('./testing.js')).auditModuleMock(importOriginal as never));
+vi.mock('@nangohq/shared', async (importOriginal) => (await import('./testing.js')).sharedModuleMock(importOriginal as never));
+
+describe('mfa audit middleware (unit)', () => {
+    beforeEach(() => {
+        installAuditMockDefaults();
+    });
+
+    afterEach(() => {
+        resetAuditMocks();
+    });
+
+    it('mfa activation: failure outcome on a rejected code, and the submitted code is never recorded', async () => {
+        const req = fakeReq({ body: { code: '000000' } });
+        const event = await runAudit(auditMfaEnabled, req, fakeRes(locals, 400));
+        expect(event).toMatchObject({
+            resource: 'mfa',
+            action: 'enabled',
+            outcome: 'failure',
+            accountId: 42,
+            // account-scoped policy → environment is never attributed.
+            environment: null,
+            actor: { type: 'user', id: '7', display: 'dev@example.com' },
+            targets: [{ type: 'user', id: '7', display: 'dev@example.com' }]
+        });
+        expect(JSON.stringify(event)).not.toContain('000000');
+    });
+});
